+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- New RPC to disable task outputs
+
 ## [0.24.0] - 2022-08-22
 
 ### Fixed